chore(refactoring): moved HA resource model to separate file
